Add the ability to track the generated object count for requests
By using the `start_processing.action_controller` event we can somewhat
track the generated object count for a request. `GC.start` is used when
we start processing the request so that we have 'fresh' count of the
objects. When the request is processed, we look again at the number of
objects and figure the difference.

Specs were added for this new feature.

README has been updated to include the new tracked feature.